+++ SBT2Blob/__init__.py
@@ -19,6 +19,24 @@ MAX_RUNTIME_SECONDS = int(os.getenv('MAX_RUNTIME_SECONDS', '0'))
 WAIT_TIME_SECONDS = int(os.getenv('WAIT_TIME_SECONDS', '5'))
 logging.basicConfig()
 logger = logging.getLogger(os.path.basename(__file__))
+_message_count = 0
+
+
+class MockTimer:
+    """
+    A mock azure.functions.TimerRequest.
+
+    Used by the main_wrapper function to call main when not part
+    of the Azure Function App tooling.
+
+    Attributes
+    ----------
+    past_due : bool
+        Always set to False.
+    """
+
+    def __init__(self):
+        self.past_due = False
 
 
 class LoadURI:
@@ -73,9 +91,12 @@ class LoadURI:
 class Extractor:
     """Extract data from Service Bus."""
 
-    def __init__(self, connection_string: str, topic_name: str, subscription_name: str):
+    def __init__(self, connection_string: str, topic_name: str, subscription_name: str,
+                 check_for_dead_letter_messages: bool):
         self.finished = False
         self.client = ServiceBusClient.from_connection_string(connection_string)
+        self.topic_name = topic_name
+        self.subscription_name = subscription_name
         self.receiver = self.client.get_subscription_receiver(
             topic_name,
             subscription_name,
@@ -83,6 +104,7 @@ class Extractor:
         )
         self.renewer = AutoLockRenewer()
         self.empty_receive_count = 0
+        self.check_for_dead_letter_messages = check_for_dead_letter_messages
 
     def accept_messages(self, messages: list[ServiceBusMessage]) -> None:
         """Accept the messages in the current buffer."""
@@ -95,6 +117,27 @@ class Extractor:
         self.receiver.close()
         self.client.close()
 
+    def dlq_has_messages(self) -> bool:
+        """
+        Check if there are dead-letter messages on the subscription.
+
+        Returns
+        -------
+        bool
+            True if there are dead-letter messages present.
+        """
+        if not self.check_for_dead_letter_messages:
+            logger.debug('Dead-letter checking is disabled.')
+            return False
+
+        msgs = self.receiver.peek_messages(max_message_count=1)
+        response = len(msgs) > 0
+
+        if response:
+            logger.warning(f'There are dead-letter messages on {self.topic_name}/{self.subscription_name}')
+
+        return response
+
     def get_messages(self) -> list[ServiceBusMessage]:
         """
         Get messages from the topic/subscription.
@@ -104,6 +147,7 @@ class Extractor:
         list
             A list of messages.
         """
+        self.dlq_has_messages()
         messages = self.receiver.receive_messages(
             max_message_count=MAX_MESSAGES_IN_BATCH,
             max_wait_time=WAIT_TIME_SECONDS
@@ -202,6 +246,7 @@ def get_environment_variable(key_name: str, default=None, required=False) -> str
 
 def main(timer: func.TimerRequest) -> None:
     """Control the main processing."""
+    global _message_count
     log_level = os.getenv('LOG_LEVEL', 'WARN')
     logger.setLevel(log_level)
     logger.debug(f'Log level is {logging.getLevelName(logger.getEffectiveLevel())}.')
@@ -216,7 +261,8 @@ def main(timer: func.TimerRequest) -> None:
     subscription_name = get_environment_variable('SUBSCRIPTION_NAME', required=True)
     topics_dir = get_environment_variable('TOPICS_DIR', default='topics')
     topic_name = get_environment_variable('TOPIC_NAME', required=True)
-    extractor = Extractor(sbns_connection_string, topic_name, subscription_name)
+    check_for_dead_letter_messages = get_environment_variable('CHECK_FOR_DL_MESSAGES', default='0') == '1'
+    extractor = Extractor(sbns_connection_string, topic_name, subscription_name, check_for_dead_letter_messages)
     loader = Loader(
         sa_connection_string,
         container_name,
@@ -224,7 +270,7 @@ def main(timer: func.TimerRequest) -> None:
         topic_name,
         path_format
     )
-    message_count = 0
+    _message_count = 0
     start_time = time.monotonic()
 
     while not extractor.finished:
@@ -238,7 +284,26 @@ def main(timer: func.TimerRequest) -> None:
         messages = extractor.get_messages()
         loader.load(messages)
         extractor.accept_messages(messages)
-        message_count += len(messages)
+        _message_count += len(messages)
 
     extractor.close()
-    logger.info(f'A total of {message_count:,} messages were loaded to blob storage for {topic_name}.')
+    logger.info(f'A total of {_message_count:,} messages were loaded to blob storage for {topic_name}.')
+
+
+def main_wrapper() -> int:
+    """
+    Call main from outside of the Azure Function App tooling.
+
+    Used by the ulti-topic-entrypoint.py script.
+
+    Returns
+    -------
+    int
+        The number of records written to file.
+    """
+    global _message_count
+
+    timer = MockTimer()
+    _message_count = 0
+    main(timer)
+    return _message_count