+++ multi-topic-entrypoint.py
@@ -3,14 +3,9 @@
 import os
 import signal
 
+from prometheus_client import Counter, start_http_server
+
 import SBT2Blob
-
-
-class MockTimer:
-    """A mock azure.functions.TimerRequest."""
-
-    def __init__(self):
-        self.past_due = False
 
 
 class Archivist:
@@ -30,14 +25,36 @@ class Archivist:
         self.topics_and_subscriptions(topics_and_subscriptions)
         signal.signal(signal.SIGINT, self.stop)
         signal.signal(signal.SIGTERM, self.stop)
+        prom_metric_prefix = os.getenv('PROMETHEUS_METRIC_NAME_PREFIX', '')
+        self.prom_files_couner = Counter(
+            f'{prom_metric_prefix}file_count',
+            'The number of files processed.'
+        )
+        self.prom_messages_counter = Counter(
+            f'{prom_metric_prefix}message_count',
+            'The number of messages processed.'
+        )
+        self.prometheus_port = int(
+            os.getenv(
+               'PROMETHEUS_PORT',
+               '8000'
+            )
+        )
 
     def run(self) -> None:
         """Run the class until a signal is received."""
+        start_http_server(self.prometheus_port)
+
         while self._is_running:
             for (topic_name, subscription_name) in self.topics_and_subscriptions():
                 os.environ['TOPIC_NAME'] = topic_name
                 os.environ['SUBSCRIPTION_NAME'] = subscription_name
-                SBT2Blob.main(MockTimer())
+                message_count = SBT2Blob.main_wrapper()
+
+                if message_count:
+                    self.prom_files_couner.inc()
+
+                self.prom_messages_counter.inc(message_count)
 
     def stop(self, signum, frame) -> None:
         """Handle a signal if received to initiate a stop."""